\name{bughunter-package}
\alias{bughunter-package}
\alias{bughunter}
\docType{package}
\title{
\packageTitle{bughunter}
}
\description{
\packageDescription{bughunter}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bughunter}
\packageIndices{bughunter}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{bughunter}

Maintainer: \packageMaintainer{bughunter}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
