% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_components.R
\name{create_console_panel}
\alias{create_console_panel}
\title{Create Console Panel UI}
\usage{
create_console_panel()
}
\value{
Shiny UI element for the console panel
}
\description{
Create Console Panel UI
}
\keyword{internal}
