% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-UI.R
\name{create_source_panel}
\alias{create_source_panel}
\title{Create Source Panel UI}
\usage{
create_source_panel()
}
\value{
Shiny UI element for the source code panel
}
\description{
Create Source Panel UI
}
\keyword{internal}
