% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-capture-accessor.R
\name{describeVariable}
\alias{describeVariable}
\title{get a description of a single variable}
\usage{
describeVariable(x, nchar = 20)
}
\description{
Describe a variable for environment display, returning a string representation
}
