% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{example_capture}
\alias{example_capture}
\title{Example Capture Object}
\format{
An object of class \code{Capture} of length 6.
}
\usage{
example_capture
}
\description{
An example Capture object obtained by running the following code:
```r
f <- function(x) {
    x <- x + 1
    g(x)
}
g <- function(y) {
    y <- y + 2
    z <- 1
    stop("Test error")
}
hunt(f(1))
example_capture <- getLastCapture()
```
}
\keyword{datasets}
