% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-TextCapture.R
\name{getEnvDescriptor,TextCapture-method}
\alias{getEnvDescriptor,TextCapture-method}
\title{Get environment descriptor for TextCapture}
\usage{
\S4method{getEnvDescriptor}{TextCapture}(capture, frameIdx)
}
\description{
TextCapture stores serialized environment information as character strings.
This method extracts the pre-saved descriptor information.
}
