% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-TextCapture.R
\name{getEnvVarDetail,TextCapture-method}
\alias{getEnvVarDetail,TextCapture-method}
\title{Get environment variable detail for TextCapture}
\usage{
\S4method{getEnvVarDetail}{TextCapture}(capture, frameIdx, varName)
}
\description{
TextCapture stores serialized variable details as character strings.
This method extracts the pre-saved detail information for a specific variable.
}
