% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLastCapture.r
\name{getLastCapture}
\alias{getLastCapture}
\title{Get the Last Stored Capture Data}
\usage{
getLastCapture(as_object = TRUE)
}
\arguments{
\item{as_object}{logical. If TRUE, returns a Capture object; if FALSE, returns a list.}
}
\value{
Either a Capture object or a list containing the capture data.
}
\description{
Retrieves the most recent data captured by hunter.
}
