% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLastHunt.r
\name{getLastHunt}
\alias{getLastHunt}
\title{Get the Last Stored Hunt Data}
\usage{
getLastHunt(as_object = TRUE)
}
\arguments{
\item{as_object}{logical. If TRUE, returns a Hunt object; if FALSE, returns a list.}
}
\value{
Either a Hunt object or a list containing the hunt data.
}
\description{
Retrieves the most recent hunt data captured by hunter.
}
