% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-environment-utils.R
\name{getVarValue}
\alias{getVarValue}
\title{get a description of a single variable}
\usage{
getVarValue(val)
}
\description{
Describe a variable for environment display, returning a string representation
}
