% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Capture.R
\name{length,Capture-method}
\alias{length,Capture-method}
\title{Length method for Capture}
\usage{
\S4method{length}{Capture}(x)
}
\arguments{
\item{x}{Capture object}
}
\description{
Returns the number of frames/calls in the capture
}
