% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hunt-class.r
\name{length,Hunt-method}
\alias{length,Hunt-method}
\title{Length method for Hunt}
\usage{
\S4method{length}{Hunt}(x)
}
\arguments{
\item{x}{Hunt object}
}
\description{
Returns the number of frames/calls in the hunt
}
