% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hunt-class.r
\name{summary,Hunt-method}
\alias{summary,Hunt-method}
\title{Summary method for Hunt}
\usage{
\S4method{summary}{Hunt}(object)
}
\arguments{
\item{object}{Hunt object}
}
\description{
Summary method for Hunt
}
