% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_logic.R
\name{trace_server}
\alias{trace_server}
\title{Server Logic for Trace Management}
\usage{
trace_server(input, output, session, trace_rv)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{trace_rv}{Reactive value containing the trace object}
}
\description{
Server Logic for Trace Management
}
\keyword{internal}
