\name{bugtracer-package}
\alias{bugtracer-package}
\alias{bugtracer}
\docType{package}
\title{
\packageTitle{bugtracer}
}
\description{
\packageDescription{bugtracer}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bugtracer}
\packageIndices{bugtracer}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{bugtracer}

Maintainer: \packageMaintainer{bugtracer}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
