% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{describeVariable}
\alias{describeVariable}
\title{get a description of a single variable}
\usage{
describeVariable(val)
}
\description{
Describe a variable for environment display, returning a string representation
}
