% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hunt.r, R/hunter.R
\name{hunt}
\alias{hunt}
\alias{hunter}
\title{Hunt Call Stacks for Error Debugging}
\usage{
hunt(expr)

hunter()
}
\arguments{
\item{expr}{An R expression to evaluate with error hunting.}

\item{condition}{The condition object (typically an error)}
}
\value{
`hunt`: results of `eval(expr)`

`hunter`: Invisible NULL
}
\description{
This function captures the call stack, deparsed code for each frame, and line numbers
when an error occurs. It stores the information in a package environment for later retrieval.
Use with options(error = hunter) to automatically capture errors.
}
