% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Capture-class.r
\name{show,Capture-method}
\alias{show,Capture-method}
\title{Show method for Capture}
\usage{
\S4method{show}{Capture}(object)
}
\arguments{
\item{object}{Capture object}
}
\description{
Show method for Capture
}
